+++ source/TestUnits/TestCode/GetLastError.pas
@@ -1,0 +1,6 @@
+program GetLastError;
+
+begin
+  WriteLn('SysErrorMessage:' + SysErrorMessage(GetLastError));
+
+end.