+++ source/TestUnits/TestCode/wd.pas
@@ -1,0 +1,6 @@
+program wd;
+
+begin
+  WriteLn('workdirectory:' + wd());
+
+end.